Homework1/requirements.txt
Jinja2
MarkupSafe
Pillow
PyWavelets
Pygments
appnope
astroid
bleach
cycler
decorator
entrypoints
html5lib
ipykernel
ipython
ipython-genutils
ipywidgets
isort
jedi
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
lazy-object-proxy
matplotlib
mccabe
mistune
nbconvert
nbformat
networkx
notebook
numpy
olefile
pandocfilters
pexpect
pickleshare
prompt-toolkit
ptyprocess
pylint
pyparsing
python-dateutil
pytz
pyzmq
qtconsole
scikit-image
scipy
simplegeneric
six
terminado
testpath
timeout-decorator
tornado
wcwidth
webencodings
widgetsnbextension
wrapt
